msm: ipa3: Fix to changes the ring variable type
With u16 variable not fitting the higher ring size. So
adding changes to chnage ring type to u32.

+++ drivers/platform/msm/gsi/gsi.h
@@ -473,7 +473,7 @@ struct gsi_chan_props {
 	uint8_t ch_id;
 	unsigned long evt_ring_hdl;
 	enum gsi_chan_ring_elem_size re_size;
-	uint16_t ring_len;
+	uint32_t ring_len;
 	uint16_t max_re_expected;
 	uint64_t ring_base_addr;
 	uint8_t db_in_bytes;
@@ -1151,7 +1151,7 @@ struct gsi_ring_ctx {
 	uint64_t rp;
 	uint64_t wp_local;
 	uint64_t rp_local;
-	uint16_t len;
+	uint32_t len;
 	uint8_t elem_sz;
 	uint16_t max_num_elem;
 	uint64_t end;
